+++ changelog.md
@@ -1,5 +1,12 @@
+# 1.0.2
+- fixed clicking esc pinning
+- mobile hold to pin
+- scrollbar for pin menu
+- button color settings
+- scale fix for sprites
+- added more ids to objects (for developers)
 # 1.0.1
-- fix english
+- fixed english
 - bug fix ordering
 - new icon
 # 1.0.0